dev: Move JoypadControlClient and JoypadControlServer as plugins
