+++ src/libYARP_dev/CMakeLists.txt
@@ -176,13 +176,6 @@ if(TARGET YARP::YARP_math)
                             src/Map2DArea.cpp)
 endif()
 
-set(YARP_dev_devices_SRCS src/devices/JoypadControlClient/JoypadControlClient.cpp
-                          src/devices/JoypadControlServer/JoypadControlServer.cpp)
-
-set(YARP_dev_devices_HDRS src/devices/JoypadControlClient/JoypadControlClient.h
-                          src/devices/JoypadControlServer/JoypadControlNetUtils.h
-                          src/devices/JoypadControlServer/JoypadControlServer.h)
-
 # Handle the YARP thrift messages
 include(YarpIDL)
 yarp_idl_to_dir(INPUT_FILES idl/stateExt.thrift
